requirements.txt
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.1
coverage==7.7.0
cryptography==44.0.2
defusedxml==0.7.1
distlib==0.3.9
dj-database-url==2.3.0
Django==5.1.7
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
djoser==2.3.1
drf-yasg==1.21.10
filelock==3.18.0
identify==2.6.9
idna==3.10
inflection==0.5.1
iniconfig==2.0.0
nodeenv==1.9.1
oauthlib==3.2.2
packaging==24.2
platformdirs==4.3.6
pluggy==1.5.0
pre_commit==4.1.0
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.9.0
pytest==8.3.5
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
social-auth-app-django==5.4.3
social-auth-core==4.5.6
sqlparse==0.5.3
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
virtualenv==20.29.3
